requirements.txt
django
django-phonenumber-field
django-environ
django-leaflet==0.32.0
psycopg2-binary==2.9.10
djangorestframework==3.16.1
markdown==3.8.2
django-filter==25.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
Pillow==11.3.0